backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.11.4
pydantic-settings==2.7.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
email-validator==2.2.0
